+++ e2e-tests/development-runtime/gatsby-node.js
@@ -77,14 +77,23 @@ exports.createPages = async function createPages({
       },
     })
   })
+
+  createPage({
+    path: `/안녕`,
+    component: path.resolve(`src/pages/page-2.js`),
+  })
+
+  createPage({
+    path: `/client-only-paths/static`,
+    component: path.resolve(`src/templates/static-page.js`),
+  })
 }
 
 exports.onCreatePage = async ({ page, actions }) => {
   const { createPage } = actions
 
-  if (page.path.match(/^\/paths/)) {
-    page.matchPath = `/paths/*`
+  if (page.path.match(/^\/client-only-paths/)) {
+    page.matchPath = `/client-only-paths/*`
+    createPage(page)
   }
-
-  createPage(page)
 }