+++ packages/gatsby/cache-dir/loader.js
@@ -1,7 +1,7 @@
 import "core-js/modules/es7.promise.finally"
 import prefetchHelper from "./prefetch"
 import emitter from "./emitter"
-import PathFinder from "./find-path"
+import { setMatchPaths, findMatchPath, cleanPath } from "./find-path"
 
 const preferDefault = m => (m && m.default) || m
 
@@ -35,19 +35,25 @@ const loadPageDataJson = loadObj => {
   return doFetch(url).then(req => {
     const { status, responseText } = req
 
-    const contentType = req.getResponseHeader(`content-type`)
-    const isJson = contentType && contentType.startsWith(`application/json`)
+    // Handle 200
+    if (status === 200) {
+      try {
+        const jsonPayload = JSON.parse(responseText)
+        if (jsonPayload.webpackCompilationHash === undefined) {
+          throw new Error(`not a valid pageData response`)
+        }
 
-    // Handle 200 JSON (Success)
-    if (status === 200 && isJson) {
-      return Object.assign(loadObj, {
-        status: `success`,
-        payload: JSON.parse(responseText),
-      })
+        return Object.assign(loadObj, {
+          status: `success`,
+          payload: jsonPayload,
+        })
+      } catch (err) {
+        // continue regardless of error
+      }
     }
 
     // Handle 404
-    if (status === 404 || (status === 200 && !isJson)) {
+    if (status === 404 || status === 200) {
       // If the request was for a 404 page and it doesn't exist, we're done
       if (pagePath === `/404.html`) {
         return Object.assign(loadObj, {
@@ -98,6 +104,7 @@ const toPageResources = (pageData, component = null) => {
     componentChunkName: pageData.componentChunkName,
     path: pageData.path,
     webpackCompilationHash: pageData.webpackCompilationHash,
+    matchPath: pageData.matchPath,
   }
 
   return {
@@ -128,7 +135,7 @@ export class BaseLoader {
     this.prefetchTriggered = new Set()
     this.prefetchCompleted = new Set()
     this.loadComponent = loadComponent
-    this.pathFinder = new PathFinder(matchPaths)
+    setMatchPaths(matchPaths)
   }
 
   setApiRunner(apiRunner) {
@@ -137,7 +144,7 @@ export class BaseLoader {
   }
 
   loadPageDataJson(rawPath) {
-    const pagePath = this.pathFinder.find(rawPath)
+    const pagePath = cleanPath(rawPath)
     if (this.pageDataDb.has(pagePath)) {
       return Promise.resolve(this.pageDataDb.get(pagePath))
     }
@@ -149,9 +156,13 @@ export class BaseLoader {
     })
   }
 
+  findMatchPath(rawPath) {
+    return findMatchPath(rawPath)
+  }
+
   // TODO check all uses of this and whether they use undefined for page resources not exist
   loadPage(rawPath) {
-    const pagePath = this.pathFinder.find(rawPath)
+    const pagePath = cleanPath(rawPath)
     if (this.pageDb.has(pagePath)) {
       const page = this.pageDb.get(pagePath)
       return Promise.resolve(page.payload)
@@ -162,6 +173,18 @@ export class BaseLoader {
 
     const inFlight = this.loadPageDataJson(pagePath)
       .then(result => {
+        if (result.notFound) {
+          // if request was a 404, we should fallback to findMatchPath.
+          let foundMatchPatch = findMatchPath(pagePath)
+          if (foundMatchPatch && foundMatchPatch !== pagePath) {
+            return this.loadPage(foundMatchPatch).then(pageResources => {
+              this.pageDb.set(pagePath, this.pageDb.get(foundMatchPatch))
+
+              return pageResources
+            })
+          }
+        }
+
         if (result.status === `error`) {
           return {
             status: `error`,
@@ -208,7 +231,7 @@ export class BaseLoader {
 
   // returns undefined if loading page ran into errors
   loadPageSync(rawPath) {
-    const pagePath = this.pathFinder.find(rawPath)
+    const pagePath = cleanPath(rawPath)
     if (this.pageDb.has(pagePath)) {
       return this.pageDb.get(pagePath).payload
     }
@@ -216,11 +239,6 @@ export class BaseLoader {
   }
 
   shouldPrefetch(pagePath) {
-    // If a plugin has disabled core prefetching, stop now.
-    if (this.prefetchDisabled) {
-      return false
-    }
-
     // Skip prefetching if we know user is on slow or constrained connection
     if (!doesConnectionSupportPrefetch()) {
       return false
@@ -238,6 +256,7 @@ export class BaseLoader {
     if (!this.shouldPrefetch(pagePath)) {
       return false
     }
+
     // Tell plugins with custom prefetching logic that they should start
     // prefetching this path.
     if (!this.prefetchTriggered.has(pagePath)) {
@@ -245,13 +264,26 @@ export class BaseLoader {
       this.prefetchTriggered.add(pagePath)
     }
 
-    this.doPrefetch(pagePath).then(pageData => {
-      const realPath = this.pathFinder.find(pagePath)
-      if (!this.prefetchCompleted.has(realPath)) {
-        this.apiRunner(`onPostPrefetchPathname`, { pathname: pagePath })
+    // If a plugin has disabled core prefetching, stop now.
+    if (this.prefetchDisabled) {
+      return false
+    }
 
-        const realPath = this.pathFinder.find(pagePath)
-        this.prefetchCompleted.add(realPath)
+    const realPath = cleanPath(pagePath)
+    // Todo make doPrefetch logic cacheable
+    // eslint-disable-next-line consistent-return
+    this.doPrefetch(realPath).then(pageData => {
+      if (!pageData) {
+        const matchPath = findMatchPath(realPath)
+
+        if (matchPath && matchPath !== realPath) {
+          return this.prefetch(matchPath)
+        }
+      }
+
+      if (!this.prefetchCompleted.has(pagePath)) {
+        this.apiRunner(`onPostPrefetchPathname`, { pathname: pagePath })
+        this.prefetchCompleted.add(pagePath)
       }
     })
 
@@ -267,7 +299,7 @@ export class BaseLoader {
   }
 
   getResourceURLsForPathname(rawPath) {
-    const pagePath = this.pathFinder.find(rawPath)
+    const pagePath = cleanPath(rawPath)
     const page = this.pageDataDb.get(pagePath)
     if (page) {
       const pageResources = toPageResources(page.payload)
@@ -282,7 +314,7 @@ export class BaseLoader {
   }
 
   isPageNotFound(rawPath) {
-    const pagePath = this.pathFinder.find(rawPath)
+    const pagePath = cleanPath(rawPath)
     const page = this.pageDb.get(pagePath)
     return page && page.notFound === true
   }