+++ packages/gatsby/src/schema/extensions/index.js
@@ -1,9 +1,6 @@
 // @flow
 const {
-  GraphQLBoolean,
-  GraphQLNonNull,
   GraphQLDirective,
-  GraphQLString,
   DirectiveLocation,
   defaultFieldResolver,
 } = require(`graphql`)
@@ -28,7 +25,7 @@ const typeExtensions = {
     description: `Infer field types from field values.`,
     args: {
       noDefaultResolvers: {
-        type: GraphQLBoolean,
+        type: `Boolean`,
         description: `Don't add default resolvers to defined fields.`,
         deprecationReason: `noDefaultResolvers is deprecated, annotate individual fields.`,
       },
@@ -38,11 +35,53 @@ const typeExtensions = {
     description: `Do not infer field types from field values.`,
     args: {
       noDefaultResolvers: {
-        type: GraphQLBoolean,
+        type: `Boolean`,
         description: `Don't add default resolvers to defined fields.`,
         deprecationReason: `noDefaultResolvers is deprecated, annotate individual fields.`,
       },
     },
+  },
+  mimeTypes: {
+    description: `Define the mime-types handled by this type.`,
+    args: {
+      types: {
+        type: `[String!]!`,
+        defaultValue: [],
+        description: `The mime-types handled by this type.`,
+      },
+    },
+  },
+  childOf: {
+    description:
+      `Define parent-child relations between types. This is used to add ` +
+      `\`child*\` or \`children*\` convenience fields like \`childImageSharp\`.`,
+    args: {
+      mimeTypes: {
+        type: `[String!]!`,
+        defaultValue: [],
+        description:
+          `A list of mime-types this type is a child of. Usually these are ` +
+          `the mime-types handled by a transformer plugin.`,
+      },
+      types: {
+        type: `[String!]!`,
+        defaultValue: [],
+        description:
+          `A list of types this type is a child of. Usually these are the ` +
+          `types handled by a transformer plugin.`,
+      },
+      many: {
+        type: `Boolean!`,
+        defaultValue: false,
+        description: `Specifies whether a parent can have multiple children of this type or not.`,
+      },
+    },
+  },
+  nodeInterface: {
+    description:
+      `Adds root query fields for an interface. All implementing types ` +
+      `must also implement the Node interface.`,
+    locations: [DirectiveLocation.INTERFACE],
   },
 }
 
@@ -51,11 +90,13 @@ const builtInFieldExtensions = {
     name: `dateformat`,
     description: `Add date formating options.`,
     args: {
-      formatString: { type: GraphQLString },
-      locale: { type: GraphQLString },
+      formatString: `String`,
+      locale: `String`,
+      fromNow: `Boolean`,
+      difference: `String`,
     },
     extend(args, fieldConfig) {
-      return getDateResolver(args)
+      return getDateResolver(args, fieldConfig)
     },
   },
 
@@ -64,16 +105,15 @@ const builtInFieldExtensions = {
     description: `Link to node by foreign-key relation.`,
     args: {
       by: {
-        type: new GraphQLNonNull(GraphQLString),
+        type: `String!`,
         defaultValue: `id`,
       },
-      from: {
-        type: GraphQLString,
-      },
+      from: `String`,
     },
     extend(args, fieldConfig) {
+      const originalResolver = fieldConfig.resolve || defaultFieldResolver
       return {
-        resolve: link(args),
+        resolve: link(args, originalResolver),
       }
     },
   },
@@ -82,13 +122,12 @@ const builtInFieldExtensions = {
     name: `fileByRelativePath`,
     description: `Link to File node by relative path.`,
     args: {
-      from: {
-        type: GraphQLString,
-      },
+      from: `String`,
     },
     extend(args, fieldConfig) {
+      const originalResolver = fieldConfig.resolve || defaultFieldResolver
       return {
-        resolve: fileByPath(args),
+        resolve: fileByPath(args, originalResolver),
       }
     },
   },
@@ -97,15 +136,13 @@ const builtInFieldExtensions = {
     name: `proxy`,
     description: `Proxy resolver from another field.`,
     args: {
-      from: {
-        type: new GraphQLNonNull(GraphQLString),
-      },
+      from: `String!`,
     },
     extend({ from }, fieldConfig) {
-      const resolver = fieldConfig.resolve || defaultFieldResolver
+      const originalResolver = fieldConfig.resolve || defaultFieldResolver
       return {
         resolve(source, args, context, info) {
-          return resolver(source, args, context, {
+          return originalResolver(source, args, context, {
             ...info,
             fieldName: from,
           })
@@ -128,17 +165,21 @@ const reservedExtensionNames = [
   ...Object.keys(builtInFieldExtensions),
 ]
 
-const toDirectives = ({ schemaComposer, extensions, locations }) =>
+const toDirectives = ({
+  schemaComposer,
+  extensions,
+  locations: defaultLocations,
+}) =>
   Object.keys(extensions).map(name => {
     const extension = extensions[name]
-    const { args, description } = extension
+    const { args, description, locations } = extension
     // Support the `graphql-compose` style of directly providing the field type as string
     const normalizedArgs = schemaComposer.typeMapper.convertArgConfigMap(args)
     return new GraphQLDirective({
       name,
       args: normalizedArgs,
       description,
-      locations,
+      locations: locations || defaultLocations,
     })
   })