+++ packages/gatsby/src/utils/webpack-utils.js
@@ -62,6 +62,7 @@ export type LoaderUtils = {
   file: LoaderResolver<*>,
   url: LoaderResolver<*>,
   js: LoaderResolver<*>,
+  dependencies: LoaderResovler<*>,
 
   miniCssExtract: LoaderResolver<*>,
   imports: LoaderResolver<*>,
@@ -255,8 +256,18 @@ module.exports = async ({
 
     js: options => {
       return {
-        options,
+        options: {
+          stage,
+          ...options,
+        },
         loader: require.resolve(`./babel-loader`),
+      }
+    },
+
+    dependencies: options => {
+      return {
+        options,
+        loader: require.resolve(`babel-loader`),
       }
     },
 
@@ -290,15 +301,35 @@ module.exports = async ({
   const rules = {}
 
   /**
-   * JavaScript loader via babel, excludes node_modules
+   * JavaScript loader via babel, includes userland code
+   * and packages that depend on `gatsby`
    */
   {
-    let js = (options = {}) => {
+    let js = (
+      { modulesThatUseGatsby, ...options } = { modulesThatUseGatsby: [] }
+    ) => {
       return {
         test: /\.(js|mjs|jsx)$/,
-        exclude: vendorRegex,
+        include: modulePath => {
+          // when it's not coming from node_modules we treat it as a source file.
+          if (!vendorRegex.test(modulePath)) {
+            return true
+          }
+
+          // If the module uses Gatsby as a dependency
+          // we want to treat it as src so we can extract queries
+          return modulesThatUseGatsby.some(module =>
+            modulePath.includes(module.path)
+          )
+        },
         type: `javascript/auto`,
-        use: [loaders.js(options)],
+        use: [
+          loaders.js({
+            ...options,
+            configFile: true,
+            compact: true,
+          }),
+        ],
       }
     }
 
@@ -306,29 +337,55 @@ module.exports = async ({
   }
 
   /**
-   * Node_modules JavaScript loader via babel (exclude core-js & babel-runtime to speedup babel transpilation)
+   * Node_modules JavaScript loader via babel
+   * Excludes core-js & babel-runtime to speedup babel transpilation
+   * Excludes modules that use Gatsby since the `rules.js` already transpiles those
    */
   {
-    let dependencies = (options = {}) => {
+    let dependencies = (
+      { modulesThatUseGatsby } = { modulesThatUseGatsby: [] }
+    ) => {
       const jsOptions = {
         babelrc: false,
         configFile: false,
         compact: false,
-        presets: [
-          [require.resolve(`babel-preset-gatsby/dependencies`), { stage }],
-        ],
+        presets: [require.resolve(`babel-preset-gatsby/dependencies`)],
         // If an error happens in a package, it's possible to be
         // because it was compiled. Thus, we don't want the browser
         // debugger to show the original code. Instead, the code
         // being evaluated would be much more helpful.
         sourceMaps: false,
+        cacheIdentifier: `${stage}---gatsby-dependencies@${
+          require(`babel-preset-gatsby/package.json`).version
+        }`,
       }
 
       return {
         test: /\.(js|mjs)$/,
-        exclude: /@babel(?:\/|\\{1,2})runtime|core-js/,
+        exclude: modulePath => {
+          if (vendorRegex.test(modulePath)) {
+            // If dep uses Gatsby, exclude
+            if (
+              modulesThatUseGatsby.some(module =>
+                modulePath.includes(module.path)
+              )
+            ) {
+              return true
+            }
+            // If dep is babel-runtime or core-js, exclude
+            if (/@babel(?:\/|\\{1,2})runtime|core-js/.test(modulePath)) {
+              return true
+            }
+
+            // If dep is in node_modules and none of the above, include
+            return false
+          }
+
+          // If dep is user land code, exclude
+          return true
+        },
         type: `javascript/auto`,
-        use: [loaders.js(jsOptions)],
+        use: [loaders.dependencies(jsOptions)],
       }
     }
 
@@ -485,7 +542,72 @@ module.exports = async ({
       ...options,
     })
 
-  plugins.minifyCss = (options = {}) => new OptimizeCssAssetsPlugin(options)
+  plugins.minifyCss = (
+    options = {
+      cssProcessorPluginOptions: {
+        preset: [
+          `default`,
+          {
+            svgo: {
+              full: true,
+              plugins: [
+                {
+                  // potentially destructive plugins removed - see https://github.com/gatsbyjs/gatsby/issues/15629
+                  // convertShapeToPath: true,
+                  // removeViewBox: true,
+                  removeUselessDefs: true,
+                  addAttributesToSVGElement: true,
+                  addClassesToSVGElement: true,
+                  cleanupAttrs: true,
+                  cleanupEnableBackground: true,
+                  cleanupIDs: true,
+                  cleanupListOfValues: true,
+                  cleanupNumericValues: true,
+                  collapseGroups: true,
+                  convertColors: true,
+                  convertPathData: true,
+                  convertStyleToAttrs: true,
+                  convertTransform: true,
+                  inlineStyles: true,
+                  mergePaths: true,
+                  minifyStyles: true,
+                  moveElemsAttrsToGroup: true,
+                  moveGroupAttrsToElems: true,
+                  prefixIds: true,
+                  removeAttributesBySelector: true,
+                  removeAttrs: true,
+                  removeComments: true,
+                  removeDesc: true,
+                  removeDimensions: true,
+                  removeDoctype: true,
+                  removeEditorsNSData: true,
+                  removeElementsByAttr: true,
+                  removeEmptyAttrs: true,
+                  removeEmptyContainers: true,
+                  removeEmptyText: true,
+                  removeHiddenElems: true,
+                  removeMetadata: true,
+                  removeNonInheritableGroupAttrs: true,
+                  removeOffCanvasPaths: true,
+                  removeRasterImages: true,
+                  removeScriptElement: true,
+                  removeStyleElement: true,
+                  removeTitle: true,
+                  removeUnknownsAndDefaults: true,
+                  removeUnusedNS: true,
+                  removeUselessStrokeAndFill: true,
+                  removeXMLNS: true,
+                  removeXMLProcInst: true,
+                  reusePaths: true,
+                  sortAttrs: true,
+                },
+              ],
+            },
+          },
+        ],
+      },
+    }
+  ) => new OptimizeCssAssetsPlugin(options)
 
   /**
    * Extracts css requires into a single file;