+++ themes/gatsby-starter-theme/gatsby-config.js
@@ -1,13 +1,13 @@
 module.exports = {
-  __experimentalThemes: [
+  plugins: [
     {
       resolve: `gatsby-theme-notes`,
       options: {
-        mdx: true,
+        mdx: false,
         basePath: `/notes`,
       },
     },
-    // with gatsby-theme-ui, the last theme in the config
+    // with gatsby-plugin-theme-ui, the last theme in the config
     // will override the theme-ui context from other themes
     { resolve: `gatsby-theme-blog` },
   ],