+++ themes/gatsby-theme-blog/gatsby-config.js
@@ -6,19 +6,17 @@ module.exports = options => {
       title: `Blog Title Placeholder`,
       author: `Name Placeholder`,
       description: `Description placeholder`,
-      siteUrl: `http://example.com/`,
       social: [
         {
-          name: `twitter`,
+          name: `Twitter`,
           url: `https://twitter.com/gatsbyjs`,
         },
         {
-          name: `github`,
+          name: `GitHub`,
           url: `https://github.com/gatsbyjs`,
         },
       ],
     },
-    __experimentalThemes: [`gatsby-theme-ui`],
     plugins: [
       mdx && {
         resolve: `gatsby-plugin-mdx`,
@@ -33,17 +31,9 @@ module.exports = options => {
                 linkImagesToOriginal: false,
               },
             },
-            { resolve: `gatsby-remark-responsive-iframe` },
             { resolve: `gatsby-remark-copy-linked-files` },
             { resolve: `gatsby-remark-numbered-footnotes` },
             { resolve: `gatsby-remark-smartypants` },
-            { resolve: `gatsby-remark-code-titles` },
-            {
-              resolve: `gatsby-remark-prismjs`,
-              options: {
-                noInlineHighlight: true,
-              },
-            },
           ],
           remarkPlugins: [require(`remark-slug`)],
         },
@@ -67,6 +57,7 @@ module.exports = options => {
       `gatsby-plugin-react-helmet`,
       `gatsby-plugin-twitter`,
       `gatsby-plugin-emotion`,
+      `gatsby-plugin-theme-ui`,
     ].filter(Boolean),
   }
 }