+++ packages/gatsby-plugin-mdx/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.23](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.22...gatsby-plugin-mdx@1.0.23) (2019-08-06)
+
+### Bug Fixes
+
+- **gatsby-plugin-mdx:** fix filter by heading depth ([#16344](https://github.com/gatsbyjs/gatsby/issues/16344)) ([234863c](https://github.com/gatsbyjs/gatsby/commit/234863c))
+
+## [1.0.22](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.21...gatsby-plugin-mdx@1.0.22) (2019-08-01)
+
+### Bug Fixes
+
+- **gatsby-plugin-mdx:** Replaces to-style with style-to-object ([#16062](https://github.com/gatsbyjs/gatsby/issues/16062)) ([a4947ef](https://github.com/gatsbyjs/gatsby/commit/a4947ef))
+
+## [1.0.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.20...gatsby-plugin-mdx@1.0.21) (2019-08-01)
+
+### Features
+
+- **docs:** Clarify gatsby-plugin-mdx requirements for gatsby-source-filesystem ([#15604](https://github.com/gatsbyjs/gatsby/issues/15604)) ([d08323c](https://github.com/gatsbyjs/gatsby/commit/d08323c))
+
+## [1.0.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.19...gatsby-plugin-mdx@1.0.20) (2019-08-01)
+
+**Note:** Version bump only for package gatsby-plugin-mdx
+
+## [1.0.19](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.18...gatsby-plugin-mdx@1.0.19) (2019-07-30)
+
+### Bug Fixes
+
+- **deps:** add missing deps to gatsby-plugin-mdx ([#16207](https://github.com/gatsbyjs/gatsby/issues/16207)) ([2db54b3](https://github.com/gatsbyjs/gatsby/commit/2db54b3))
+
+## [1.0.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.17...gatsby-plugin-mdx@1.0.18) (2019-07-29)
+
+**Note:** Version bump only for package gatsby-plugin-mdx
+
+## [1.0.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.16...gatsby-plugin-mdx@1.0.17) (2019-07-25)
+
+### Bug Fixes
+
+- **gatsby-plugin-mdx:** Add missing HTML attribute translations ([#16043](https://github.com/gatsbyjs/gatsby/issues/16043)) ([36e742a](https://github.com/gatsbyjs/gatsby/commit/36e742a))
+
+## [1.0.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.15...gatsby-plugin-mdx@1.0.16) (2019-07-23)
+
+### Bug Fixes
+
+- **gatsby-plugin-mdx:** pass maxDepth correctly to mdast-util-toc ([#15994](https://github.com/gatsbyjs/gatsby/issues/15994)) ([129b141](https://github.com/gatsbyjs/gatsby/commit/129b141)), closes [/github.com/syntax-tree/mdast-util-toc/blob/beca120e03db0a715b8517248968bf34290ebd1a/lib/search.js#L18](https://github.com//github.com/syntax-tree/mdast-util-toc/blob/beca120e03db0a715b8517248968bf34290ebd1a/lib/search.js/issues/L18)
+
+## [1.0.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.14...gatsby-plugin-mdx@1.0.15) (2019-07-20)
+
+### Bug Fixes
+
+- **gatsby-plugin-mdx:** Fix graphql query examples in readme ([#15691](https://github.com/gatsbyjs/gatsby/issues/15691)) ([c64d322](https://github.com/gatsbyjs/gatsby/commit/c64d322))
+
+## [1.0.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.13...gatsby-plugin-mdx@1.0.14) (2019-07-19)
+
+**Note:** Version bump only for package gatsby-plugin-mdx
+
+## [1.0.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.12...gatsby-plugin-mdx@1.0.13) (2019-07-13)
+
+**Note:** Version bump only for package gatsby-plugin-mdx
+
+## [1.0.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.11...gatsby-plugin-mdx@1.0.12) (2019-07-12)
+
+### Bug Fixes
+
+- **gatsby-plugin-mdx:** Fix Safari 9 `const` syntax error ([#15644](https://github.com/gatsbyjs/gatsby/issues/15644)) ([a823463](https://github.com/gatsbyjs/gatsby/commit/a823463))
+
+## [1.0.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.10...gatsby-plugin-mdx@1.0.11) (2019-07-11)
+
+### Bug Fixes
+
+- **gatsby-plugin-mdx:** add missing change-case dependency ([#15638](https://github.com/gatsbyjs/gatsby/issues/15638)) ([74d33c6](https://github.com/gatsbyjs/gatsby/commit/74d33c6))
+
+## [1.0.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.9...gatsby-plugin-mdx@1.0.10) (2019-07-11)
+
+**Note:** Version bump only for package gatsby-plugin-mdx
+
+## [1.0.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.8...gatsby-plugin-mdx@1.0.9) (2019-07-10)
+
+**Note:** Version bump only for package gatsby-plugin-mdx
+
+## [1.0.8](https://github.com/johno/gatsby/compare/gatsby-plugin-mdx@1.0.7...gatsby-plugin-mdx@1.0.8) (2019-07-09)
+
+### Bug Fixes
+
+- **gatsby-plugin-mdx:** Use babel plugin to remove export keyword ([#15452](https://github.com/johno/gatsby/issues/15452)) ([d3582b0](https://github.com/johno/gatsby/commit/d3582b0)), closes [ChristopherBiscardi/gatsby-mdx#411](https://github.com/ChristopherBiscardi/gatsby-mdx/issues/411) [/github.com/mdx-js/mdx#618](https://github.com//github.com/mdx-js/mdx/issues/618)
+
+## [1.0.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.6...gatsby-plugin-mdx@1.0.7) (2019-07-02)
+
+**Note:** Version bump only for package gatsby-plugin-mdx
+
 ## 1.0.6 (2019-06-28)
 
 ### Bug Fixes