+++ packages/gatsby-remark-images-contentful/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-images-contentful@2.1.3...gatsby-remark-images-contentful@2.1.4) (2019-07-13)
+
+**Note:** Version bump only for package gatsby-remark-images-contentful
+
+## [2.1.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-images-contentful@2.1.2...gatsby-remark-images-contentful@2.1.3) (2019-07-12)
+
+**Note:** Version bump only for package gatsby-remark-images-contentful
+
+## [2.1.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-images-contentful@2.1.1...gatsby-remark-images-contentful@2.1.2) (2019-07-11)
+
+**Note:** Version bump only for package gatsby-remark-images-contentful
+
+## [2.1.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-images-contentful@2.1.0...gatsby-remark-images-contentful@2.1.1) (2019-07-02)
+
+**Note:** Version bump only for package gatsby-remark-images-contentful
+
 # [2.1.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-images-contentful@2.0.13...gatsby-remark-images-contentful@2.1.0) (2019-06-20)
 
 **Note:** Version bump only for package gatsby-remark-images-contentful