+++ packages/gatsby-source-wikipedia/CHANGELOG.md
@@ -3,15 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [2.1.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wikipedia/compare/gatsby-source-wikipedia@2.0.8...gatsby-source-wikipedia@2.1.0) (2019-06-20)
+## [2.1.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-wikipedia@2.1.1...gatsby-source-wikipedia@2.1.2) (2019-07-12)
+
+### Bug Fixes
+
+- correct links in package changelogs ([#15630](https://github.com/gatsbyjs/gatsby/issues/15630)) ([d07b9dd](https://github.com/gatsbyjs/gatsby/commit/d07b9dd))
+
+## [2.1.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-wikipedia@2.1.0...gatsby-source-wikipedia@2.1.1) (2019-07-11)
 
 **Note:** Version bump only for package gatsby-source-wikipedia
 
-## [2.0.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wikipedia/compare/gatsby-source-wikipedia@2.0.7...gatsby-source-wikipedia@2.0.8) (2019-05-31)
+# [2.1.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-wikipedia@2.0.8...gatsby-source-wikipedia@2.1.0) (2019-06-20)
 
 **Note:** Version bump only for package gatsby-source-wikipedia
 
-## [2.0.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-wikipedia/compare/gatsby-source-wikipedia@2.0.6...gatsby-source-wikipedia@2.0.7) (2019-04-23)
+## [2.0.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-wikipedia@2.0.7...gatsby-source-wikipedia@2.0.8) (2019-05-31)
+
+**Note:** Version bump only for package gatsby-source-wikipedia
+
+## [2.0.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-wikipedia@2.0.6...gatsby-source-wikipedia@2.0.7) (2019-04-23)
 
 **Note:** Version bump only for package gatsby-source-wikipedia